requirements.txt
fastapi
uvicorn
supabase
google-generativeai
pandas
python-dotenv